clippy fixes

+++ libzhuyin/src/parser.rs
@@ -233,7 +233,7 @@ impl ZhuyinParser {
                             if alt.chars().count() == substr.chars().count() {
                                 let end = pos + len;
                                 let seg_cost =
-                                    (penalty as f32) * (config.fuzzy_penalty_multiplier as f32); // use fuzzy penalty from config
+                                    penalty * (config.fuzzy_penalty_multiplier as f32); // use fuzzy penalty from config
                                 let cand = seg_cost + dp_cost[end];
                                 if cand < dp_cost[pos] {
                                     dp_cost[pos] = cand;
